🚚rename: 디렉토리 구조를 도메인 중심으로 수정
* 이유: 비즈니스 로직이 많아지면서 같은 서비스의 각종 레이어들을 한번에 보기 힘들어져 한 패키지에 모아둠
