+++ src/main/java/kr/codesqaud/cafe/user/UserRepository.java
@@ -1,8 +1,7 @@
-package kr.codesqaud.cafe.repository;
+package kr.codesqaud.cafe.user;
 
 import java.util.List;
 import java.util.Optional;
-import kr.codesqaud.cafe.domain.User;
 
 public interface UserRepository {
