+++ src/main/java/kr/codesqaud/cafe/user/UserService.java
@@ -1,11 +1,8 @@
-package kr.codesqaud.cafe.service;
+package kr.codesqaud.cafe.user;
 
 import java.util.List;
 import java.util.Optional;
-import kr.codesqaud.cafe.domain.User;
-import kr.codesqaud.cafe.dto.UserDTO;
 import kr.codesqaud.cafe.exception.signUpException.InvalidUserIdException;
-import kr.codesqaud.cafe.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
 /**
