+++ src/test/java/kr/codesqaud/cafe/repository/UserRepositoryImplTest.java
@@ -2,7 +2,9 @@ package kr.codesqaud.cafe.repository;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-import kr.codesqaud.cafe.domain.User;
+import kr.codesqaud.cafe.user.User;
+import kr.codesqaud.cafe.user.UserRepository;
+import kr.codesqaud.cafe.user.UserRepositoryImpl;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
