+++ src/test/java/kr/codesqaud/cafe/service/ArticleServiceTest.java
@@ -2,12 +2,14 @@ package kr.codesqaud.cafe.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import kr.codesqaud.cafe.domain.Article;
-import kr.codesqaud.cafe.dto.UserDTO;
-import kr.codesqaud.cafe.repository.ArticleRepository;
-import kr.codesqaud.cafe.repository.ArticleRepositoryImpl;
-import kr.codesqaud.cafe.repository.UserRepository;
-import kr.codesqaud.cafe.repository.UserRepositoryImpl;
+import kr.codesqaud.cafe.article.Article;
+import kr.codesqaud.cafe.article.ArticleRepository;
+import kr.codesqaud.cafe.article.ArticleRepositoryImpl;
+import kr.codesqaud.cafe.article.ArticleService;
+import kr.codesqaud.cafe.user.UserDTO;
+import kr.codesqaud.cafe.user.UserRepository;
+import kr.codesqaud.cafe.user.UserRepositoryImpl;
+import kr.codesqaud.cafe.user.UserService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
