+++ src/test/java/kr/codesqaud/cafe/service/LoginServiceTest.java
@@ -2,10 +2,11 @@ package kr.codesqaud.cafe.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import kr.codesqaud.cafe.domain.User;
-import kr.codesqaud.cafe.dto.LoginFormDTO;
-import kr.codesqaud.cafe.repository.UserRepository;
-import kr.codesqaud.cafe.repository.UserRepositoryImpl;
+import kr.codesqaud.cafe.login.LoginFormDTO;
+import kr.codesqaud.cafe.login.LoginService;
+import kr.codesqaud.cafe.user.User;
+import kr.codesqaud.cafe.user.UserRepository;
+import kr.codesqaud.cafe.user.UserRepositoryImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
